# openapi.json
{"openapi":"3.0.2","info":{"title":"ResearchGPT","description":"ResearchGPT is a tool for researchers to generate ideas and insights.","version":"0.1.0"},"paths":{"/":{"get":{"summary":"Root","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/chatroom/":{"get":{"summary":"Get Chatrooms","operationId":"get_chatrooms_chatroom__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatRoom"}}}}}},"post":{"summary":"Create Chatroom","operationId":"create_chatroom_chatroom__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatRoomCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatRoom"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/chatroom/{chatroom_id}":{"put":{"summary":"Update Chatroom","operationId":"update_chatroom_chatroom__chatroom_id__put","parameters":[{"required":true,"schema":{"title":"Chatroom Id","type":"integer"},"name":"chatroom_id","in":"path"}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatRoomUpdate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatRoom"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"summary":"Delete Chatroom","operationId":"delete_chatroom_chatroom__chatroom_id__delete","parameters":[{"required":true,"schema":{"title":"Chatroom Id","type":"integer"},"name":"chatroom_id","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatRoom"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/chatroom/{chatroom_id}/{last_message_id}":{"get":{"summary":"Get Chatroom","operationId":"get_chatroom_chatroom__chatroom_id___last_message_id__get","parameters":[{"required":true,"schema":{"title":"Chatroom Id","type":"integer"},"name":"chatroom_id","in":"path"},{"required":true,"schema":{"title":"Last Message Id","type":"integer"},"name":"last_message_id","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatRoomMessage"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"ChatRoom":{"title":"ChatRoom","required":["id","user_id","title","search"],"type":"object","properties":{"id":{"title":"Id","type":"integer"},"user_id":{"title":"User Id","type":"integer"},"title":{"title":"Title","type":"string"},"search":{"title":"Search","type":"string"}}},"ChatRoomCreate":{"title":"ChatRoomCreate","required":["search"],"type":"object","properties":{"search":{"title":"Search","type":"string"}}},"ChatRoomMessage":{"title":"ChatRoomMessage","required":["id","chatroom_id","user_id","message"],"type":"object","properties":{"id":{"title":"Id","type":"integer"},"chatroom_id":{"title":"Chatroom Id","type":"integer"},"user_id":{"title":"User Id","type":"integer"},"message":{"title":"Message","type":"string"}}},"ChatRoomUpdate":{"title":"ChatRoomUpdate","required":["title"],"type":"object","properties":{"title":{"title":"Title","type":"string"}}},"HTTPValidationError":{"title":"HTTPValidationError","type":"object","properties":{"detail":{"title":"Detail","type":"array","items":{"$ref":"#/components/schemas/ValidationError"}}}},"ValidationError":{"title":"ValidationError","required":["loc","msg","type"],"type":"object","properties":{"loc":{"title":"Location","type":"array","items":{"anyOf":[{"type":"string"},{"type":"integer"}]}},"msg":{"title":"Message","type":"string"},"type":{"title":"Error Type","type":"string"}}}}}}